Move packagemanager wrapper into dagger layer
