+++ src/main/java/com/pyamsoft/padlock/dagger/wrapper/PackageManagerWrapper.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.pyamsoft.padlock.app.wrapper;
+package com.pyamsoft.padlock.dagger.wrapper;
 
 import android.content.pm.ActivityInfo;
 import android.content.pm.ApplicationInfo;
